@@ -3,7 +3,7 @@ from __future__ import annotations
 import copy
 import math
 import random
-from typing import Iterable
+from collections.abc import Iterable
 
 import numpy as np
 import torch
