@@ -2,7 +2,8 @@ from __future__ import annotations
 
 import random
 import warnings
-from typing import Any, Iterable
+from collections.abc import Iterable
+from typing import Any
 
 import torch.nn.functional as F
 from torch import Tensor, nn
