@@ -7,7 +7,7 @@ from __future__ import annotations
 import csv
 import gzip
 import os
-from typing import Generator
+from collections.abc import Generator
 
 import pytest
 import torch
