@@ -7,8 +7,8 @@ from __future__ import annotations
 import csv
 import gzip
 import os
+from collections.abc import Generator
 from pathlib import Path
-from typing import Generator
 
 import numpy as np
 import pytest
