@@ -8,7 +8,7 @@ class SafeTemporaryDirectory(tempfile.TemporaryDirectory):
     The GitHub Actions CI on Windows sometimes raises a NotADirectoryError when cleaning up the temporary directory.
     This class is a workaround to avoid the error.
 
-    Unlike tempfile.TemporaryDirectory(ignore_cleanup_errors=True), this also works on Python 3.8 and 3.9.
+    Unlike tempfile.TemporaryDirectory(ignore_cleanup_errors=True), this also works on Python 3.9.
     """
 
     def __init__(self, *args, **kwargs) -> None:
